// extensions/openhue/src/api/generated/src/models/LightGet.ts
export interface LightGet {
  /**
   * Type of the supported resources
   * @type {string}
   * @memberof LightGet
   */
  type?: string;
  /**
   * Unique identifier representing a specific resource instance
   * @type {string}
   * @memberof LightGet
   */
  id?: string;
  /**
   * Clip v1 resource identifier
   * @type {string}
   * @memberof LightGet
   */
  id_v1?: string;
  /**
   *
   * @type {ResourceIdentifier}
   * @memberof LightGet
   */
  owner?: ResourceIdentifier;
  /**
   *
   * @type {LightGetAllOfMetadata}
   * @memberof LightGet
   */
  metadata?: LightGetAllOfMetadata;
  /**
   *
   * @type {On}
   * @memberof LightGet
   */
  on?: On;
  /**
   *
   * @type {LightGetAllOfDimming}
   * @memberof LightGet
   */
  dimming?: LightGetAllOfDimming;
  /**
   *
   * @type {LightGetAllOfColorTemperature}
   * @memberof LightGet
   */
  color_temperature?: LightGetAllOfColorTemperature;
  /**
   *
   * @type {LightGetAllOfColor}
   * @memberof LightGet
   */
  color?: LightGetAllOfColor;
  /**
   *
   * @type {LightGetAllOfDynamics}
   * @memberof LightGet
   */
  dynamics?: LightGetAllOfDynamics;
  /**
   * TODO
   * @type {object}
   * @memberof LightGet
   */
  alert?: object;
  /**
   *
   * @type {LightGetAllOfSignaling}
   * @memberof LightGet
   */
  signaling?: LightGetAllOfSignaling;
  /**
   *
   * @type {string}
   * @memberof LightGet
   */
  mode?: LightGetModeEnum;
  /**
   *
   * @type {object}
   * @memberof LightGet
   */
  gradient?: object;
  /**
   *
   * @type {LightGetAllOfEffects}
   * @memberof LightGet
   */
  effects?: LightGetAllOfEffects;
  /**
   *
   * @type {LightGetAllOfTimedEffects}
   * @memberof LightGet
   */
  timed_effects?: LightGetAllOfTimedEffects;
  /**
   *
   * @type {LightGetAllOfPowerup}
   * @memberof LightGet
   */
  powerup?: LightGetAllOfPowerup;
}

/**
 * @export
 */
export const LightGetModeEnum = {
  normal: "normal",
  streaming: "streaming",
} as const;
export type LightGetModeEnum = (typeof LightGetModeEnum)[keyof typeof LightGetModeEnum];