// extensions/openhue/src/api/generated/src/models/RoomGet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hue CLIP API
 * See [openhue/openhue-api](https://github.com/openhue/openhue-api)
 *
 * The version of the OpenAPI document: v2
 * Contact: contact@openhue.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { ResourceIdentifier } from "./ResourceIdentifier";
import {
  ResourceIdentifierFromJSON,
  ResourceIdentifierFromJSONTyped,
  ResourceIdentifierToJSON,
  ResourceIdentifierToJSONTyped,
} from "./ResourceIdentifier";
import type { RoomGetAllOfMetadata } from "./RoomGetAllOfMetadata";
import {
  RoomGetAllOfMetadataFromJSON,
  RoomGetAllOfMetadataFromJSONTyped,
  RoomGetAllOfMetadataToJSON,
  RoomGetAllOfMetadataToJSONTyped,
} from "./RoomGetAllOfMetadata";

/**
 *
 * @export
 * @interface RoomGet
 */
export interface RoomGet {
  /**
   * Type of the supported resources
   * @type {string}
   * @memberof RoomGet
   */
  type?: string;
  /**
   * Unique identifier representing a specific resource instance
   * @type {string}
   * @memberof RoomGet
   */
  id?: string;
  /**
   * Clip v1 resource identifier
   * @type {string}
   * @memberof RoomGet
   */
  id_v1?: string;
  /**
   * Child devices/services to group by the derived group
   * @type {Array<ResourceIdentifier>}
   * @memberof RoomGet
   */
  children?: Array<ResourceIdentifier>;
  /**
   * References all services aggregating control and state of children in the group.
   * This includes all services grouped in the group hierarchy given by child relation.
   * This includes all services of a device grouped in the group hierarchy given by child relation.
   * Aggregation is per service type, ie every service type which can be grouped has a corresponding definition of
   * grouped type.
   * Supported types: – grouped_light
   *
   * @type {Array<ResourceIdentifier>}
   * @memberof RoomGet
   */
  services?: Array<ResourceIdentifier>;
  /**
   *
   * @type {RoomGetAllOfMetadata}
   * @memberof RoomGet
   */
  metadata?: RoomGetAllOfMetadata;
}

/**
 * Check if a given object implements the RoomGet interface.
 */
export function instanceOfRoomGet(value: object): value is RoomGet {
  return true;
}

export function RoomGetFromJSON(json: any): RoomGet {
  return RoomGetFromJSONTyped(json, false);
}

export function RoomGetFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoomGet {
  if (json == null) {
    return json;
  }
  return {
    type: json["type"] == null ? undefined : json["type"],
    id: json["id"] == null ? undefined : json["id"],
    id_v1: json["id_v1"] == null ? undefined : json["id_v1"],
    children: json["children"] == null ? undefined : (json["children"] as Array<any>).map(ResourceIdentifierFromJSON),
    services: json["services"] == null ? undefined : (json["services"] as Array<any>).map(ResourceIdentifierFromJSON),
    metadata: json["metadata"] == null ? undefined : RoomGetAllOfMetadataFromJSON(json["metadata"]),
  };
}

export function RoomGetToJSON(json: any): RoomGet {
  return RoomGetToJSONTyped(json, false);
}

export function RoomGetToJSONTyped(value?: RoomGet | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    type: value["type"],
    id: value["id"],
    id_v1: value["id_v1"],
    children: value["children"] == null ? undefined : (value["children"] as Array<any>).map(ResourceIdentifierToJSON),
    services: value["services"] == null ? undefined : (value["services"] as Array<any>).map(ResourceIdentifierToJSON),
    metadata: RoomGetAllOfMetadataToJSON(value["metadata"]),
  };
}